requirements.txt
fastapi
uvicorn
pydantic
openai==0.27.0
requests
httpx
tiktoken
Pillow
telethon